+++ cmd/metal-api/internal/testdata/ipam.go
@@ -42,6 +42,7 @@ func InitMockIpamData(dbMock *r.Mock, withIP bool) (ipam.IPAMer, error) {
 		if err != nil {
 			return nil, fmt.Errorf("error creating ipam mock data: %w", err)
 		}
+		fmt.Printf("Created IP:%s\n", ipAddress)
 		IPAMIP.IPAddress = ipAddress
 		IPAMIP.ParentPrefixCidr = prefixesIPAM[0].String()
 		IPAMIP.NetworkID = NwIPAM.ID
